feat: 修复日历/照片墙/地图/个人资料等多项功能
- 修复日历1、2号不显示（ForEach ID冲突）
- 照片墙统一正方形网格排版
- 草稿箱增加清空按钮
- 新增拍照+相册选择功能
- 修复ISO8601日期解析兼容性
- 个人资料编辑保存（头像+昵称）
- 删除账号确认弹窗
- 地图返回地球按钮修复
- 地图pin详情跳转打卡详情页
- 后端UUID类型转换修复500错误
- 后端Decimal改float修复iOS解码失败

+++ backend/app/schemas/checkin.py
@@ -48,7 +48,7 @@ class CheckInResponse(BaseModel):
     tags: list[str]
     note: str | None
     is_public: bool
-    amount: Decimal | None
+    amount: float | None
     amount_type: str | None
     created_at: datetime
